friend adapter

+++ Clockio/app/src/main/java/com/natasha/clockio/base/model/PageResponse.kt
@@ -5,10 +5,10 @@ data class PageResponse<T> (
     val first: Boolean,
     val last: Boolean,
     val empty: Boolean,
-    val totalPages: Long,
-    val totalElements: Long,
-    val numberOfElements: Long,
-    val size: Long, // page size
-    val number: Long // page number
+    val totalPages: Int,
+    val totalElements: Int,
+    val numberOfElements: Int,
+    val size: Int, // page size
+    val number: Int // page number
 
 )
